requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
seaborn>=0.13.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
python-dateutil>=2.8.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-experimental>=0.0.47
fpdf2>=2.7.0
tabulate>=0.9.0
statsmodels>=0.14.0
